requirements.txt
pandas
rank_bm25
nltk
openai
python-dotenv
loguru
transformers